pyproject.toml
[tool.poetry]
name = "chatapp"
version = "0.1.0"
description = ""
authors = ["dedoxd2 <dedo.xd33@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
